ODPresentation Reader : Read differents units for margin

+++ src/PhpPresentation/Shape/RichText/Paragraph.php
@@ -72,12 +72,12 @@ class Paragraph implements ComparableInterface
     private $lineSpacingMode = self::LINE_SPACING_MODE_PERCENT;
 
     /**
-     * @var int
+     * @var float
      */
     private $spacingBefore = 0;
 
     /**
-     * @var int
+     * @var float
      */
     private $spacingAfter = 0;
 
@@ -346,7 +346,7 @@ class Paragraph implements ComparableInterface
     /**
      * Value in points.
      */
-    public function getSpacingBefore(): int
+    public function getSpacingBefore(): float
     {
         return $this->spacingBefore;
     }
@@ -354,7 +354,7 @@ class Paragraph implements ComparableInterface
     /**
      * Value in points.
      */
-    public function setSpacingBefore(int $spacingBefore): self
+    public function setSpacingBefore(float $spacingBefore): self
     {
         $this->spacingBefore = $spacingBefore;
 
@@ -364,7 +364,7 @@ class Paragraph implements ComparableInterface
     /**
      * Value in points.
      */
-    public function getSpacingAfter(): int
+    public function getSpacingAfter(): float
     {
         return $this->spacingAfter;
     }
@@ -372,7 +372,7 @@ class Paragraph implements ComparableInterface
     /**
      * Value in points.
      */
-    public function setSpacingAfter(int $spacingAfter): self
+    public function setSpacingAfter(float $spacingAfter): self
     {
         $this->spacingAfter = $spacingAfter;
